clean up campaign-manage
