+++ frontend/forms/__init__.py
@@ -4,7 +4,7 @@ import logging
 import re
 import unicodedata
 
-from datetime import timedelta, date
+from datetime import date
 from decimal import Decimal as D
 
 #django imports
@@ -15,8 +15,6 @@ from django.contrib.auth.models import User
 from django.forms.widgets import RadioSelect
 from django.forms.extras.widgets import SelectDateWidget
 from django.utils.translation import ugettext_lazy as _
-
-from ckeditor.widgets import CKEditorWidget
 
 from selectable.forms import (
     AutoCompleteSelectWidget,
@@ -57,14 +55,17 @@ from regluit.core.lookups import (
     SubjectLookup,
 )
 from regluit.core.validation import test_file
-from regluit.utils.localdatetime import now
 from regluit.utils.fields import ISBNField
 
 
 from .bibforms import EditionForm, IdentifierForm
 from .rh_forms import (
+    CCDateForm,
     CloneCampaignForm,
+    date_selector,
+    DateCalculatorForm,
     EditManagersForm,
+    ManageCampaignForm,
     OpenCampaignForm,
     RightsHolderForm,
     UserClaimForm
@@ -288,140 +289,6 @@ class OfferForm(forms.ModelForm):
                 'license': forms.HiddenInput,
             }
 
-date_selector = range(date.today().year, settings.MAX_CC_DATE.year+1)
-
-class CCDateForm(object):
-    target = forms.DecimalField(
-        min_value= D(settings.UNGLUEIT_MINIMUM_TARGET),
-        error_messages={'required': 'Please specify a Revenue Target.'}
-    )
-    minimum_target = settings.UNGLUEIT_MINIMUM_TARGET
-    maximum_target = settings.UNGLUEIT_MAXIMUM_TARGET
-    max_cc_date = settings.MAX_CC_DATE
-
-    def clean_target(self):
-        new_target = self.cleaned_data['target']
-        if new_target < D(settings.UNGLUEIT_MINIMUM_TARGET):
-            raise forms.ValidationError(_('A campaign may not be launched with a target less than $%s' % settings.UNGLUEIT_MINIMUM_TARGET))
-        if new_target > D(settings.UNGLUEIT_MAXIMUM_TARGET):
-            raise forms.ValidationError(_('A campaign may not be launched with a target more than $%s' % settings.UNGLUEIT_MAXIMUM_TARGET))
-        return new_target
-
-    def clean_cc_date_initial(self):
-        new_cc_date_initial = self.cleaned_data['cc_date_initial']
-        if new_cc_date_initial.date() > settings.MAX_CC_DATE:
-            raise forms.ValidationError('The initial Ungluing Date cannot be after %s'%settings.MAX_CC_DATE)
-        elif new_cc_date_initial - now() < timedelta(days=0):
-            raise forms.ValidationError('The initial Ungluing date must be in the future!')
-        return new_cc_date_initial
-
-class DateCalculatorForm(CCDateForm, forms.ModelForm):
-    revenue = forms.DecimalField()
-    cc_date_initial = forms.DateTimeField(
-            widget = SelectDateWidget(years=date_selector)
-        )
-    class Meta:
-        model = Campaign
-        fields = 'target',  'cc_date_initial', 'revenue',
-
-def getManageCampaignForm ( instance, data=None, initial=None, *args, **kwargs ):
-
-    def get_queryset():
-        work = instance.work
-        return Edition.objects.filter(work = work)
-
-    class ManageCampaignForm(CCDateForm, forms.ModelForm):
-        target = forms.DecimalField( required= (instance.type in {REWARDS, BUY2UNGLUE}))
-        deadline = forms.DateTimeField(
-                required = (instance.type==REWARDS),
-                widget = SelectDateWidget(years=date_selector) if instance.status=='INITIALIZED' else forms.HiddenInput
-            )
-        cc_date_initial = forms.DateTimeField(
-                required = (instance.type==BUY2UNGLUE) and instance.status=='INITIALIZED',
-                widget = SelectDateWidget(years=date_selector) if instance.status=='INITIALIZED' else forms.HiddenInput
-            )
-        paypal_receiver = forms.EmailField(
-            label=_("contact email address for this campaign"),
-            max_length=100,
-            error_messages={'required': 'You must enter the email we should contact you at for this campaign.'},
-            )
-        edition = forms.ModelChoiceField(
-            get_queryset(),
-            widget=RadioSelect(),
-            empty_label='no edition selected',
-            required=False,
-        )
-        publisher = forms.ModelChoiceField(
-            instance.work.publishers(),
-            empty_label='no publisher selected',
-            required=False,
-        )
-        work_description =  forms.CharField( required=False , widget=CKEditorWidget())
-
-        class Meta:
-            model = Campaign
-            fields = ('description', 'details', 'license', 'target', 'deadline', 'paypal_receiver',
-                'edition', 'email', 'publisher',  'cc_date_initial', "do_watermark", "use_add_ask",
-            )
-            widgets = { 'deadline': SelectDateWidget }
-
-        def clean_target(self):
-            if self.instance.type == THANKS:
-                return None
-            new_target = super(ManageCampaignForm, self).clean_target()
-            if self.instance:
-                if self.instance.status == 'ACTIVE' and self.instance.target < new_target:
-                    raise forms.ValidationError(_('The fundraising target for an ACTIVE campaign cannot be increased.'))
-            return new_target
-
-        def clean_cc_date_initial(self):
-            if self.instance.type in {REWARDS, THANKS} :
-                return None
-            if self.instance:
-                if self.instance.status != 'INITIALIZED':
-                    # can't change this once launched
-                    return self.instance.cc_date_initial
-            return super(ManageCampaignForm, self).clean_cc_date_initial()
-
-        def clean_deadline(self):
-            if self.instance.type in {BUY2UNGLUE, THANKS} :
-                return None
-            new_deadline_date = self.cleaned_data['deadline']
-            new_deadline = new_deadline_date + timedelta(hours=23, minutes=59)
-            if self.instance:
-                if self.instance.status == 'ACTIVE':
-                    return self.instance.deadline
-            if new_deadline_date - now() > timedelta(days=int(settings.UNGLUEIT_LONGEST_DEADLINE)):
-                raise forms.ValidationError(_('The chosen closing date is more than %s days from now' % settings.UNGLUEIT_LONGEST_DEADLINE))
-            elif new_deadline - now() < timedelta(days=0):
-                raise forms.ValidationError(_('The chosen closing date is in the past'))
-            return new_deadline
-
-        def clean_license(self):
-            new_license = self.cleaned_data['license']
-            if self.instance:
-                if self.instance.status == 'ACTIVE' and self.instance.license != new_license:
-                    # should only allow change to a less restrictive license
-                    if self.instance.license == 'CC BY-ND' and new_license in ['CC BY-NC-ND', 'CC BY-NC-SA', 'CC BY-NC']:
-                        raise forms.ValidationError(_('The proposed license for an ACTIVE campaign may not add restrictions.'))
-                    elif self.instance.license == 'CC BY' and new_license != 'CC0':
-                        raise forms.ValidationError(_('The proposed license for an ACTIVE campaign may not add restrictions.'))
-                    elif self.instance.license == 'CC BY-NC' and new_license in ['CC BY-NC-ND', 'CC BY-NC-SA', 'CC BY-SA', 'CC BY-ND']:
-                        raise forms.ValidationError(_('The proposed license for an ACTIVE campaign may not add restrictions.'))
-                    elif self.instance.license == 'CC BY-ND' and new_license in ['CC BY-NC-ND', 'CC BY-NC-SA', 'CC BY-SA', 'CC BY-NC']:
-                        raise forms.ValidationError(_('The proposed license for an ACTIVE campaign may not add restrictions.'))
-                    elif self.instance.license == 'CC BY-SA' and new_license in ['CC BY-NC-ND', 'CC BY-NC-SA', 'CC BY-ND', 'CC BY-NC']:
-                        raise forms.ValidationError(_('The proposed license for an ACTIVE campaign may not add restrictions.'))
-                    elif self.instance.license == 'CC BY-NC-SA' and new_license in ['CC BY-NC-ND', 'CC BY-ND']:
-                        raise forms.ValidationError(_('The proposed license for an ACTIVE campaign may not add restrictions.'))
-                    elif self.instance.license == 'CC0' :
-                        raise forms.ValidationError(_('The proposed license for an ACTIVE campaign may not add restrictions.'))
-                    elif self.instance.license in ['GDFL', 'LAL']:
-                        raise forms.ValidationError(_('Once you start a campaign with GDFL or LAL, you can\'t use any other license.'))
-            return new_license
-    if initial and not initial.get('edition', None) and not instance.edition:
-        initial['edition'] = instance.work.editions.all()[0]
-    return ManageCampaignForm(instance=instance, data=data, initial=initial)
 
 class CampaignPurchaseForm(forms.Form):
     anonymous = forms.BooleanField(required=False, label=_("Make this purchase anonymous, please"))
